pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gemini-live-telemetry"
version = "0.1.0"
description = "Instrumentation and telemetry for the Google Gemini Live API via OpenTelemetry"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "AjnaLens" },
]
keywords = [
    "gemini", "google", "genai", "live-api", "telemetry",
    "opentelemetry", "monitoring", "metrics", "instrumentation",
    "vertex-ai", "voice", "realtime",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "wrapt>=2.1.0",
    "google-genai>=1.5.0",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-gcp-monitoring>=1.11.0a0",
    "google-cloud-monitoring-dashboards>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.urls]
Homepage = "https://github.com/AjnaLens/gemini-live-telemetry"
Repository = "https://github.com/AjnaLens/gemini-live-telemetry"
Issues = "https://github.com/AjnaLens/gemini-live-telemetry/issues"
Documentation = "https://github.com/AjnaLens/gemini-live-telemetry#readme"

[project.scripts]
gemini-live-compare = "gemini_live_telemetry.compare:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_live_telemetry"]
